#75 Conform protocol to C-R1.1
Model depends on data module (not the opposite)
The controller relies more on model (and not on data module)
Add toJsonForClient method into model structure
Add icon into Challenge class

+++ domain/src/main/java/fr/unice/polytech/ecoknowledge/domain/model/conditions/basic/CounterDeserializer.java
@@ -15,7 +15,7 @@ public class CounterDeserializer extends JsonDeserializer<Counter> {
 			throws IOException, JsonProcessingException {
 		JsonNode node = jp.getCodec().readTree(jp);
 		String type = (node.get("type")).asText();
-		Double threshold = (node.get("threshold")).asDouble();
+		Integer threshold = (node.get("threshold")).asInt();
 
 		return new Counter(threshold, CounterType.fromString(type));
 	}
